+++ src/test/java/com/expedia/www/haystack/commons/secretDetector/xml/XmlS3ConfigFetcherTest.java
@@ -251,7 +251,6 @@ public class XmlS3ConfigFetcherTest {
         assertNotNull(factory.createBufferedReader(mockInputStreamReader));
     }
 
-    @SuppressWarnings("LawOfDemeter")
     @Test
     public void testFactoryCreateWhiteListItem() {
         final XmlWhiteListItem whiteListItem = factory.createWhiteListItem(